Have pip force-reinstall packages always
Pip, if it finds an existing package, doesn't install it again even if we've specified `--root`/`--prefix`. This was discovered when we updated the version of `six` to be the same as latest. Komodo has `six` as one of its dependencies and so pip discovered that it was already installed in the temporary Komodo virtualenv, and so didn't install it into the release.