Store hash instead of branch to release manifest
This will store the full commit hash of the cloned ref to the release
manifest, instead of the tag/branch name. A minimal git repo is
added to the test data to allow for testing of a known hash in
the integration test. This also adds testing of fetching using
git clone in the integration test, which was previously not tested.