Don't have `testkomodo.sh` automatically run code
This breaks all current `testkomodo.sh` scripts, but it should make them
much simpler in the long run. `testkomodo.sh` scripts can expect
`ci_helper.sh` functions to be available with the prefix `ci_`, but
other than that it is up to each and every script to handle the setup
and testing. `run_tests` is no longer a function that is ran.