Create named arg group and improve default
There is no default for the pip cache, resulting
in an ungraceful error if no cache is provided.
This change adds the default `pip-cache` in the
CLI & removes the downstream default in build.py.

Note that the default has changed (it was `.`),
but not really because using the default resulted
in an error.

By default argparse's 'optional' group includes all named
arguments, even if they are required. That's confusing.

This change adds a required named argument group with the
3 other required arguments. We have to suppress the automatic
addition of the --help option and add it explcitly to an
optional group.