modernize build system
Proof of concept for moving komodo to PEP 621 packaging, which
uses only `pyproject.toml` (no `setup.py` or `setup.cfg`) and
is meant to be easier to maintain. For example, all build tool
options can go into the TOML file. (One minor exception: pylint
doesn't support it yet, so max-line-length option was added
to command-line call in style.yml GitHub  workflow.)

This changes how the project is built, requiring the `build`
module. The wheel and sdist are produced by `python -m build`.
And `pip install .` still works.

Note that the dynamic versioning depends on git tags and requires
a `komodo/_version.py` file **which should not be kept under
version control**. Added it to .gitignore.