remove locations-config from required args
locations-config, which sets up the local/local.csh scripts,
is likely not required by non-Equinor users of komodo and
should not be a required argument. Should not need deprecating
because it can only be used explicitly currently.

Also updated the README because the example in there can now
be shorter and cleaner. And removed the explicit cache as well
because that was defaulted in the "Group kmd arguments" PR.

Finally, added a 'minimal' which uses the example from the
README and checks that one can indeed call kmd without the
locations-config argument (and also without the cache,
which we made optional in an earlier PR).