Deprecate pyver in the CLI
The pyver option is present in the CLI but there is no code
that uses args.pyver so the option does nothing. Can't just
remove it because code using it will break. So this commit
adds a conspicuous warning (which admittedly gets a bit
buried by subsequent output). We can remove the option in a
few weeks.

Tried adding a DeprecationWarning, but while this is raised
and can be tested for, it doesn't show in the CLI <shrug>.

Includes a test to check that the deprecation message appears
in stdout.