Fix yaml packagename casing validation
Recent yaml file type commit added validation to make sure packages had lowercase names.
This was a bad idea, because we are linting against old release files, and we do not want
to update those. Instead, make it so that it checks for typing consistencies between the files,
and output a warning if the specific package is not found, but one with a different casing exists.